pyproject.toml
[tool.poetry]
name = "async-doc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
fastapi = "^0.103.1"
python = "^3.11"
twilio = "^8.9.0"
openai = "^0.28.0"
python-decouple = "^3.8"
langchain = "^0.0.300"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.1"
sqlmodel = "^0.0.8"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-multipart = "^0.0.6"
python-docx = "^0.8.11"
boto3 = "^1.28.53"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.291"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
